+++ ObjectPrintingTests/ObjectPrinterAcceptanceTests.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Globalization;
 using NUnit.Framework;
+using ObjectPrinting;
 using ObjectPrinting.Extensions;
 
-namespace ObjectPrinting.Tests
+namespace ObjectPrintingTests
 {
     [TestFixture]
     public class ObjectPrinterAcceptanceTests
