+++ ObjectPrintingTests/ObjectPrinterTests.cs
@@ -1,10 +1,10 @@
-﻿using FluentAssertions;
+﻿using System.Collections.Generic;
+using System.Globalization;
+using FluentAssertions;
 using NUnit.Framework;
 using ObjectPrinting.Extensions;
-using System.Collections.Generic;
-using System.Globalization;
 
-namespace ObjectPrinting.Tests
+namespace ObjectPrintingTests
 {
     public class ObjectPrinterTests
     {
